Metabase 0.45.1 upgrade
0.42.0 > 0.45.1

This upgrade takes Metabase to the same version we are running for the
Report environment.

We have run into problems with the service in Canada running v0.42.0.
The service in the US is working as expected so this is a bit strange.
Various steps to troubleshoot that included, redeploying the existing
version and comparing the US and CA authentication configurations have
not solved the problem.

We were expecting Google to make a change to their authentication
services at the end of March, but I am wondering if something has change
in Canada ahead of that date...